+++ src/GafferScene/SceneAlgo.cpp
@@ -36,6 +36,7 @@
 
 #include "GafferScene/SceneAlgo.h"
 
+#include "GafferScene/AttributeTweaks.h"
 #include "GafferScene/CameraTweaks.h"
 #include "GafferScene/CopyAttributes.h"
 #include "GafferScene/Filter.h"
@@ -588,7 +589,7 @@ void addShuffleAttributesPredecessors( const ShuffleAttributes *shuffleAttribute
 	destination->predecessors.push_back( SceneAlgo::attributeHistory( source[0].get(), sourceAttributeName ) );
 }
 
-void addLocaliseAttributesPredecessors( const LocaliseAttributes *localiseAttributes, const SceneAlgo::History::Predecessors &source, SceneAlgo::AttributeHistory *destination )
+void addLocaliseAttributesPredecessors( const SceneAlgo::History::Predecessors &source, SceneAlgo::AttributeHistory *destination )
 {
 	// No need to check if the node is filtered to this location.
 	// Filtered or unfiltered, it's all the same : the predecessor
@@ -611,7 +612,11 @@ void addLocaliseAttributesPredecessors( const LocaliseAttributes *localiseAttrib
 		}
 	}
 
-	assert( predecessor );
+	if( !predecessor )
+	{
+		return;
+	}
+
 	destination->predecessors.push_back( predecessor );
 }
 
@@ -759,13 +764,17 @@ SceneAlgo::AttributeHistory::Ptr SceneAlgo::attributeHistory( const SceneAlgo::H
 		{
 			addShuffleAttributesPredecessors( shuffleAttributes, attributesHistory->predecessors, result.get() );
 		}
-		else if( auto localiseAttributes = runTimeCast<const LocaliseAttributes>( node ) )
+		else if( runTimeCast<const LocaliseAttributes>( node ) )
 		{
-			addLocaliseAttributesPredecessors( localiseAttributes, attributesHistory->predecessors, result.get() );
+			addLocaliseAttributesPredecessors( attributesHistory->predecessors, result.get() );
 		}
 		else if( auto mergeScenes = runTimeCast<const MergeScenes>( node ) )
 		{
 			addMergeScenesPredecessors( mergeScenes, attributesHistory->predecessors, result.get() );
+		}
+		else if( runTimeCast<const AttributeTweaks>( node ) )
+		{
+			addLocaliseAttributesPredecessors( attributesHistory->predecessors, result.get() );
 		}
 		else
 		{